refactor(task-session-manager): extract continuation-token + input-wait managers
Split two cohesive subsystems out of the 1,355-line index.ts into their own
modules:
- continuation-token-manager.ts: session-token lifecycle (get/is-current/
  invalidate/clear) with onInvalidateContinuation callback for idle cleanup.
- input-wait-tracker.ts: permission/question wait tracking with injected
  invalidateContinuation callback.

Pure extraction; behavior and exports unchanged. index.ts drops ~93 lines.
Includes incidental Biome formatting fixes in adjacent test/tracker files.
